apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
neo4j>=5.16.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
GitPython>=3.1.41
sqlglot>=20.0.0